moviehub/requirements.txt
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
click==8.1.3
decorator==5.1.1
executing==1.2.0
Flask==2.2.2
grpcio==1.51.1
grpcio-tools==1.51.1
ipython==8.7.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.33
protobuf==4.21.12
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
six==1.16.0
stack-data==0.6.2
traitlets==5.6.0
wcwidth==0.2.5
Werkzeug==2.2.2
pytest==6.2.5